[python] Use a custom modtime implementation
It only exposes "monotonic" and "sleep" functions, which are the
only part of CPython's time we can actually do at the moment.
